JSCBC-436: Updated to libcouchbase 2.8.3.
Reviewed-on: http://review.couchbase.org/86399
Tested-by: Brett Lawson <brett19@gmail.com>

+++ deps/lcb/src/n1ql/params.cc
@@ -1,3 +1,20 @@
+/* -*- Mode: C; tab-width: 4; c-basic-offset: 4; indent-tabs-mode: nil -*- */
+/*
+ *     Copyright 2015-2017 Couchbase, Inc.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
 #include <libcouchbase/couchbase.h>
 #include <libcouchbase/n1ql.h>
 #include <libcouchbase/vbucket.h>
@@ -120,7 +137,7 @@ encode_mutation_token(Json::Value& sparse, const lcb_MUTATION_TOKEN *sv)
     Json::Value& cur_sv = sparse[buf];
 
     cur_sv[0] = static_cast<Json::UInt64>(sv->seqno_);
-    sprintf(buf, "%llu", sv->uuid_);
+    sprintf(buf, "%llu", (unsigned long long)sv->uuid_);
     cur_sv[1] = buf;
 }
